Creates HomScreen component and adds NavigationContainer in App.jsx
