+++ App.jsx
@@ -1,40 +1,33 @@
-import { StatusBar } from 'expo-status-bar';
 import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
-import { useSelector } from 'react-redux'
+import { useSelector } from 'react-redux';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
+import HomeScreen from './components/HomeScreen';
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'left',
-    justifyContent: 'center',
-  },
-});
-
-const HomeScreen = () => {
-  return (
-    <View style={styles.container}>
-      <ApplicationHeader />
-      <Text>{appTitle}</Text>
-      <StatusBar style="light" />
-    </View>
-  )
-}
-
-const Stack = createStackNavigator()
+const Stack = createStackNavigator();
 
 const App = () => {
-  const { appTitle } = useSelector(state => state)
+  const { appTitle } = useSelector((state) => state);
   return (
-  <NavigationContainer>
+    <NavigationContainer>
+      <Stack.Navigator>
+        <Stack.Screen
+          name='Home'
+          component={HomeScreen}
+          options={{
+            title: appTitle,
+            headerStyle: {
+              backgroundColor: 'teal',
+            },
+            headerTitleStyle: {
+              color: 'white',
+              fontSize: 30,
+            },
+          }}
+        />
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+};
 
-  </NavigationContainer>
-  )
-}
-
-
-
-export default App
+export default App;
